backend/requirements.txt
# Kingly backend – FastAPI + ADK + Supabase
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
google-adk[extensions]>=0.1.0
supabase>=2.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
google-generativeai>=0.8.0
python-multipart>=0.0.9
pyjwt>=2.8.0
httpx>=0.27.0
jwcrypto>=1.5.0
